python/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==4.2.3
django-environ==0.10.0
djangorestframework==3.14.0
exceptiongroup==1.1.2
falu==1.0.1
idna==3.4
iniconfig==2.0.0
packaging==23.1
pluggy==1.2.0
pytest==7.4.0
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
responses==0.23.1
sqlparse==0.4.4
tomli==2.0.1
types-PyYAML==6.0.12.11
typing_extensions==4.7.1
urllib3==1.26.14
